WinML/Transformers/GoogleBert/requirements.txt
--index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/ORT-Nightly/pypi/simple
--extra-index-url https://pypi.org/simple

numpy
transformers
wasdk-Microsoft.Windows.AI.MachineLearning[all]>=2.0.0.dev1
wasdk-Microsoft.Windows.ApplicationModel.DynamicDependency.Bootstrap[all]>=2.0.0.dev1
onnxruntime-winml
torch
onnxscript

